@@ -1,14 +1,16 @@
 package com.ismasurance.ismasurance.user.api.v1.dto;
 
 import lombok.Data;
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotBlank;
 
 
 @Data
 public class UserRequest {
-    @NotBlank(message= "userName")
+    @NotBlank(message= "userId is required")
+    private String userId;
+    @NotBlank(message= "userName is required")
     private String userName;
-    @NotBlank(message= "password")
+    @NotBlank(message= "password is required")
     private String password;
     private String email;
     private String phone;
