@@ -4,9 +4,7 @@ import com.ismasurance.ismasurance.user.data.UserEntity;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 @Repository
 public interface UserRepository extends MongoRepository<UserEntity, String> {
-    UserEntity findUserByUserName(String email);
+    UserEntity findByEmail(String email);
 }
